modified scripts ot make onts
